+++ src/headless/headless-events.ts
