+++ src/tests/auto-mode-piped.test.ts
@@ -93,7 +93,7 @@ export async function resolve(specifier, context, nextResolve) {
   if (specifier === '@gsd/pi-ai' || specifier === '@gsd/pi-ai/oauth') return { url: 'stub:pi-ai', shortCircuit: true }
   if (specifier === '@gsd/pi-tui') return { url: 'stub:pi-tui', shortCircuit: true }
   if (specifier === 'chalk') return { url: 'stub:chalk', shortCircuit: true }
-  if (specifier === './headless.js' && context.parentURL?.endsWith('/src/cli.ts')) {
+  if (specifier === './headless/headless.js' && context.parentURL?.endsWith('/src/cli.ts')) {
     return { url: 'stub:headless', shortCircuit: true }
   }
   if (
