+++ tsconfig.test.json
@@ -7,8 +7,8 @@
   "include": [
     "src/tests/headless-cli-surface.test.ts",
     "src/tests/ensure-workspace-builds.test.ts",
-    "src/headless-events.ts",
-    "src/headless-types.ts",
+    "src/headless/headless-events.ts",
+    "src/headless/headless-types.ts",
     "src/tests/google-search-oauth-shape.test.ts",
     "src/tests/google-search-auth.repro.test.ts"
   ],
